+++ test/sql/type.sql
@@ -3,6 +3,9 @@
 SET client_min_messages = warning;
 SET search_path=datalink;
 
+select dlvalue(null);
+select dlvalue('');
+select dlvalue('foo');
 select dlvalue('http://www.ljudmila.org/');
 select dlvalue('http://www.ljudmila.org/','URL','Example datalink');
 select dlvalue('/tmp','FS','Example file datalink');