+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Version 0.12
 - now uses attnum instead of attname in dl_linked_files
 - file_path domain used in args to file functions
 - alter table rename column works now
+- dlvalue(null) returns null, per SQL spec
 
 Version 0.11
 ------------